Profile page cleanup, prelim styling (#115)
Thanks Kevin for this!

* Some initial cleanup

- JS -> TS
- Self-closing elements
- Some `var` removal, state assignment removal

This still doesn't compile because of lack of typing. My word!

* Remove unnecessary fragments, more state / ref cleanup

* Abstract auto-resizing textareas to own component

* Remove unnecessary "lock type" derived state

* Better profile data typing, inline fetch function

* Remove derived bio state, unnecessary `if` nesting

- `userData.bio && userData.bio != ''` is a redundant check since `''` is falsy and will fail the first condition always.

* Remove unnecessary `profileData` state

* Remove media icon derived states

* Remove media icon classname derived states

Relatively wasteful, as derived states go.

* Remove over-cautious checks

If you guys control the shape of this object, then there wouldn't be a reason that fields like `bio` or `userPfp` are missing from the object, right? If they are meant to be nullable, it should be typed as so.

* Remove "profile contact" derived state

* Remove unneeded refs

* Remove final unneeded ref & derived state

* Remove unnecessary wrapper divs, better textarea styling

Use `disabled:` to conditionally style the textarea when editing is disabled rather than the previous "store refs and empirically update style values" model.

* Make pfp circular, better edit styling w/ hover, remove black border

This unfortunately loses image optimization, but I don't know how to get `next/image` to clip the image to 200x200.

* Basic focus styling